+++ core/src/main/java/com/ray3k/particleparkpro/Listeners.java
@@ -7,7 +7,6 @@ import com.badlogic.gdx.scenes.scene2d.*;
 import com.badlogic.gdx.scenes.scene2d.actions.Actions;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
-import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener.ChangeEvent;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.Disableable;
 import com.badlogic.gdx.scenes.scene2d.utils.FocusListener;
@@ -20,6 +19,7 @@ import com.ray3k.stripe.ScrollFocusListener;
 import com.ray3k.stripe.Spinner;
 
 import static com.ray3k.particleparkpro.Core.*;
+import static com.ray3k.particleparkpro.PreviewSettings.*;
 import static com.ray3k.particleparkpro.widgets.styles.Styles.infSliderStyle;
 import static com.ray3k.particleparkpro.widgets.styles.Styles.tooltipBottomRightArrowStyle;
 
@@ -166,8 +166,9 @@ public class Listeners {
 
             @Override
             public void enter(InputEvent event, float x, float y, int pointer, Actor fromActor) {
-                var pressingButtons = Gdx.input.isButtonPressed(Buttons.LEFT) || Gdx.input.isButtonPressed(Buttons.RIGHT) || Gdx.input.isButtonPressed(Buttons.MIDDLE);
-                if (pointer == -1 && !pressingButtons && popTable.isHidden() && !dismissed) {
+                if (!Utils.isWindowFocused()) return;
+                if (Gdx.input.isButtonPressed(Buttons.LEFT) || Gdx.input.isButtonPressed(Buttons.RIGHT) || Gdx.input.isButtonPressed(Buttons.MIDDLE)) return;
+                if (pointer == -1 && popTable.isHidden() && !dismissed) {
                     if (fromActor == null || !event.getListenerActor().isAscendantOf(fromActor)) {
                         if (showTableAction == null) {
                             showTableAction = Actions.delay(.5f,
@@ -233,8 +234,8 @@ public class Listeners {
         sliderPop.attachToActor(valueSpinner, Align.bottom, Align.bottom);
 
         var slider = new InfSlider(infSliderStyle);
-        slider.setRange(adjustByPPM ? range / ParticlePreview.pixelsPerMeter : range);
-        slider.setIncrement(adjustByPPM ? increment / ParticlePreview.pixelsPerMeter : increment);
+        slider.setRange(adjustByPPM ? range / getPixelsPerMeter() : range);
+        slider.setIncrement(adjustByPPM ? increment / getPixelsPerMeter() : increment);
         slider.addListener(noCaptureKeyboardFocusListener);
         slider.getKnob().addListener(noCaptureKeyboardFocusListener);
         slider.getBackground().addListener(noCaptureKeyboardFocusListener);