+++ core/src/main/java/com/ray3k/particleparkpro/widgets/subpanels/ImagesSubPanel.java
@@ -166,7 +166,7 @@ public class ImagesSubPanel extends Panel {
 
             @Override
             public void selected(String text) {
-
+                updateDisabled();
             }
         });